+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2019-02-12
 ### Added
 - Initial setup with working composer install.
+- New Service contract, DB schema setup and basic preloading of db tables
 
 ### Changed
 - Initial code repurposed, moved from a different repository to this one